requirements.txt
pyquaternion==0.9.9
contourpy==1.3.2
widgetsnbextension==4.0.14
tensorboard==2.17.1
mpmath==1.3.0
safetensors==0.5.3
nvidia-cusolver-cu11==11.4.1.48
blinker==1.9.0
portalocker==3.2.0
asttokens==3.0.0
wheel==0.45.1
huggingface-hub==0.34.3
platformdirs==4.3.8
joblib==1.5.1
iopath==0.1.10
Markdown==3.8
jsonschema-specifications==2025.9.1
plotly==6.3.1
qwen-vl-utils==0.0.11
pycocotools==2.0.10
dash==3.2.0
addict==2.4.0
wandb==0.21.0
ptyprocess==0.7.0
Flask==3.1.2
six==1.17.0
parso==0.8.5
matplotlib-inline==0.1.7
torchvision==0.20.0+cu118
frozenlist==1.6.0
OpenEXR==3.4.1
click==8.2.1
cycler==0.12.1
aiosignal==1.3.2
tzdata==2025.2
pexpect==4.9.0
torch==2.5.0+cu118
sentencepiece==0.1.99
zipp==3.23.0
multiprocess==0.70.16
nvidia-cublas-cu11==11.11.3.6
idna==3.10
certifi==2025.4.26
nest-asyncio==1.6.0
open3d==0.19.0
KNN-CUDA==0.2
multidict==6.4.4
rpds-py==0.27.1
jsonschema==4.25.1
python-dateutil==2.9.0.post0
aiohappyeyeballs==2.6.1
nvidia-nccl-cu11==2.21.5
fsspec==2025.3.0
nvidia-ml-py==12.575.51
pandas==2.2.3
exceptiongroup==1.3.0
ninja==1.11.1.4
matplotlib==3.9.2
sentry-sdk==2.34.1
torchaudio==2.5.0+cu118
ipywidgets==8.1.7
gitdb==4.0.12
annotated-types==0.7.0
lit==18.1.8
executing==2.2.1
cmake==4.0.2
packaging==25.0
deepspeed==0.17.1
fastjsonschema==2.21.2
absl-py==2.1.0
charset-normalizer==3.4.2
pip==25.1
plyfile==1.1
tensorboard-data-server==0.7.2
prompt_toolkit==3.0.52
nvidia-cudnn-cu11==9.1.0.70
networkx==3.4.2
jedi==0.19.2
pointnet2_ops==3.0.0
drjit==1.2.0
pyparsing==3.2.3
GitPython==3.1.45
imath==0.0.2
ConfigArgParse==1.7.1
nvidia-curand-cu11==10.3.0.86
tqdm==4.67.1
hf-xet==1.1.5
stack-data==0.6.3
av==15.0.0
py-cpuinfo==9.0.0
transformers==4.54.1
psutil==7.0.0
pytorch3d==0.7.8
Pygments==2.19.2
pydantic==2.11.7
Jinja2==3.1.6
scipy==1.14.1
protobuf==6.31.0
urllib3==2.4.0
nvidia-cuda-nvrtc-cu11==11.8.89
itsdangerous==2.2.0
pyarrow==20.0.0
comm==0.2.3
sympy==1.13.1
einops==0.6.1
importlib_metadata==8.7.0
filelock==3.18.0
nvidia-cufft-cu11==10.9.0.58
MarkupSafe==3.0.2
triton==3.1.0
nvidia-cusparse-cu11==11.7.5.86
hjson==3.1.0
nvidia-cuda-cupti-cu11==11.8.87
traitlets==5.14.3
pillow==11.2.1
dill==0.3.8
Werkzeug==3.1.3
nbformat==5.10.4
jupyterlab_widgets==3.0.15
regex==2024.11.6
peft==0.17.0
narwhals==2.7.0
nvidia-nvtx-cu11==11.8.86
mitsuba==3.7.1
scikit-learn==1.6.1
mpi4py==4.1.0
timm==0.9.12
PyYAML==6.0.2
wcwidth==0.2.14
typing-inspection==0.4.1
retrying==1.4.2
trl==0.20.0
propcache==0.3.1
setuptools==78.1.1
pydantic_core==2.33.2
jupyter_core==5.8.1
termcolor==2.4.0
xxhash==3.5.0
typing_extensions==4.13.2
referencing==0.36.2
grpcio==1.71.0
smmap==5.0.2
yarl==1.20.0
accelerate==1.7.0
msgpack==1.1.1
pure_eval==0.2.3
easydict==1.13
attrs==25.3.0
decorator==5.2.1
threadpoolctl==3.6.0
pytz==2025.2
requests==2.32.3
kiwisolver==1.4.8
ipython==8.37.0
async-timeout==5.0.1
aiohttp==3.12.2
flash-attn==2.7.3
datasets==3.6.0
numpy==1.26.4
fonttools==4.58.0
opencv-python==4.11.0.86
tokenizers==0.21.2
nvidia-cuda-runtime-cu11==11.8.89
autocommand==2.2.2
wheel==0.45.1
typing_extensions==4.12.2
backports.tarfile==1.2.0
jaraco.functools==4.0.1
jaraco.collections==5.1.0
tomli==2.0.1
packaging==24.2
typeguard==4.3.0
more-itertools==10.3.0
platformdirs==4.2.2
jaraco.context==5.3.0
importlib_metadata==8.0.0
zipp==3.19.2
inflect==7.3.1
jaraco.text==3.12.1